+++ DoctorBeaver/HeaderCell.swift
@@ -9,15 +9,15 @@
 import UIKit
 
 class HeaderCell: UITableViewCell {
-
+  
   @IBOutlet weak var titleLabel: UILabel!
   
-    override func awakeFromNib() {
-        super.awakeFromNib()
-    }
-
-    override func setSelected(selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-    }
-
+  override func awakeFromNib() {
+    super.awakeFromNib()
+  }
+  
+  override func setSelected(selected: Bool, animated: Bool) {
+    super.setSelected(selected, animated: animated)
+  }
+  
 }
