+++ DoctorBeaver/MenuDataPickerCell.swift
@@ -1,17 +1,17 @@
 //
-//  SettingDescriptionCell.swift
+//  MenuDataPickerCell.swift
 //  DoctorBeaver
 //
-//  Created by Anastasia Stepanova-Kolupakhina on 18.02.16.
+//  Created by Anastasia Stepanova-Kolupakhina on 19.02.16.
 //  Copyright © 2016 Anastasia Stepanova-Kolupakhina. All rights reserved.
 //
 
 import UIKit
 
-class StgTitleValueCell: UITableViewCell {
+class MenuDataPickerCell: UITableViewCell {
   
-  @IBOutlet weak var titleLabel: UILabel!
-  @IBOutlet weak var valueLabel: UILabel!
+  @IBOutlet weak var dataPickerView: DataPickerView!
+  
   override func awakeFromNib() {
     super.awakeFromNib()
   }
@@ -19,4 +19,5 @@ class StgTitleValueCell: UITableViewCell {
   override func setSelected(selected: Bool, animated: Bool) {
     super.setSelected(selected, animated: animated)
   }
+  
 }
