--- DoctorBeaver/Pet+CoreDataProperties.swift
@@ -1,15 +1,0 @@
-//
-//  Pet+CoreDataProperties.swift
-//  DoctorBeaver
-//
-//  Created by Anastasia Stepanova-Kolupakhina on 16.03.16.
-//  Copyright © 2016 Anastasia Stepanova-Kolupakhina. All rights reserved.
-//
-//  Choose "Create NSManagedObject Subclass…" from the Core Data editor menu
-//  to delete and recreate this implementation file for your updated model.
-//
-
-import Foundation
-import CoreData
-
-
