+++ DoctorBeaver/String.swift
@@ -22,5 +22,23 @@ extension String {
     }
   }
   
+  static func getOneDimArrayOfStrings(fromUnseparatedString string: String, withSeparator separator: Character) -> [String] {
+    let oneDimArray = string.characters.split(separator, maxSplit: string.characters.count, allowEmptySlices: true).map{String($0)}
+    return oneDimArray
+  }
+  
+  static func getTwoDimArrayOfStrings(fromUnseparatedString string: String, withSeparator separator: Character) -> [[String]] {
+    
+    let twoDimSeparator = String(separator) + String(separator)
+    var twoDimArray = [[String]]()
+    
+    let twoDimStringElements = string.componentsSeparatedByString(twoDimSeparator)
+    for twoDimStringElement in twoDimStringElements {
+      twoDimArray.append(getOneDimArrayOfStrings(fromUnseparatedString: twoDimStringElement, withSeparator: separator))
+    }
+    //return twoDimArray
+    return (twoDimArray.filter{ !$0.isEmpty }).isEmpty ? [] : twoDimArray
+  }
+  
 }
