+++ DoctorBeaver/TaskTypeItemBasicValues.swift
@@ -9,6 +9,21 @@
 import Foundation
 import CoreData
 
+extension TaskTypeItemBasicValues {
+  @NSManaged var typeItem: NSSet
+  
+  @NSManaged var taskNamePlaceholder: String
+  @NSManaged var separator: String
+  
+  @NSManaged var startDateTitle: String
+  @NSManaged var daysOptions: String
+  @NSManaged var endDaysOrTimesTitle: String
+  @NSManaged var endDaysOrTimesSegmentTitles: String
+  @NSManaged var endDaysOrTimesOptionsPreposition: String
+  @NSManaged var timesOptions: String
+  
+  @NSManaged var commentPlaceholder: String
+}
 
 class TaskTypeItemBasicValues: NSManagedObject {
